Add trsv, add timing, refactor, fix errors
Details:
- Add trsv and tests (preliminary)
- Add first of timing tests to be used for comparison to NumPy (gemm)
- Refactor tests by moving everything up a level
- Fixed an error in the tester on some versions of NumPy with not using the intended dtype
- Fixed an error in the errors.py file which did not list all values in a tuple when throwing an error

+++ bp_tests/helpers.py
@@ -44,4 +44,4 @@ def random_triangular_matrix(n, dtype, as_matrix, uplo, diag, trans='n'):
     if diag == 'u' or diag == 'U':
         fill_diagonal(rand_matrix, 1)
     rand_matrix = triu(rand_matrix) if (uplo == 'u' or uplo == 'U') else tril(rand_matrix)
-    return rand_matrix if (trans == 'n' or trans == 'N') else rand_matrix.T
+    return (rand_matrix if (trans == 'n' or trans == 'N') else rand_matrix.T).astype(dtype)
